chore: refactor error into dedicated file
