+++ worker/src/features/evaluation/evalService.ts
@@ -56,29 +56,7 @@ import { compileHandlebarString, createW3CTraceId } from "../utils";
 import { env } from "../../env";
 import { JSONPath } from "jsonpath-plus";
 import { UnrecoverableError } from "../../errors/UnrecoverableError";
-
-const ObservationNotFoundErrorName = "ObservationNotFoundError";
-
-export class ObservationNotFoundError extends Error {
-  observationId: string;
-
-  constructor(params: { message: string; observationId: string }) {
-    super(params.message);
-
-    this.name = ObservationNotFoundErrorName;
-    this.observationId = params.observationId;
-
-    if (Error.captureStackTrace) {
-      Error.captureStackTrace(this);
-    }
-  }
-}
-
-export function isObservationNotFoundError(
-  e: any,
-): e is ObservationNotFoundError {
-  return e instanceof Error && e.name === ObservationNotFoundErrorName;
-}
+import { ObservationNotFoundError } from "../../errors/ObservationNotFoundError";
 
 let s3StorageServiceClient: StorageService;
